Expand simulation by adding controller
Add controller which task is to control
the process variable (temperature) of control object
(Room) by specifying the value of control variable
(power level of the heater).
Implement two types of controller: bang-bang
(on-of) in BBController and PID controller in
PIDController which can be used in simulation through
interface ControllerIf.