Add abstract classes for Room and Heater classes
Add abstract class ControlObjectIf which Room
class derrives from, and Actuator if class
which Heater class derrives from.
This refactoring improves testability and
opens a possibility to create simulations
with different scenarios in the future.

Change the composition of Room and Heater
in the Simulation to aggregation. Use
contructor DI, to remove scrict coupling.